requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Redis
redis==5.0.7
hiredis==2.3.2

# HTTP Client for OpenRouter
httpx==0.27.0

# Agent & LLM
openai==1.35.7  # OpenRouter uses OpenAI compatible API
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Logging
structlog==24.2.0

# Utils
python-json-logger==2.0.7
tenacity==8.4.2  # For retry logic

# Web Search
duckduckgo-search==6.1.7  # Free web search API